# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Postcard API",
    "description": "Digital forensics API for verifying social media posts.",
    "version": "1.0.0",
    "contact": {
      "name": "Postcard HQ",
      "url": "https://github.com/postcardhq/postcard"
    }
  },
  "servers": [
    {
      "url": "https://postcard.fartlabs.org/api",
      "description": "Production API"
    },
    {
      "url": "/api",
      "description": "Local/Relative API base"
    }
  ],
  "paths": {
    "/postcards": {
      "get": {
        "summary": "Retrieve existing analysis",
        "description": "Retrieves the forensic verification result for a given URL if it exists in the cache.",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "description": "The social media post URL to verify",
            "schema": {
              "type": "string",
              "format": "uri"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Analysis found or processing",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": { "type": "string", "format": "uri" },
                    "postcardId": { "type": "string", "format": "uuid" },
                    "status": {
                      "type": "string",
                      "enum": ["processing", "completed", "failed"]
                    },
                    "platform": { "type": "string" },
                    "postcardScore": {
                      "type": "number",
                      "description": "Postcard score (0-100)"
                    },
                    "markdown": { "type": "string" },
                    "progress": { "type": "number" },
                    "stage": { "type": "string" },
                    "message": { "type": "string" },
                    "error": { "type": "string" },
                    "postcard": {
                      "type": "object",
                      "properties": {
                        "platform": {
                          "type": "string",
                          "enum": [
                            "X",
                            "YouTube",
                            "Reddit",
                            "Instagram",
                            "Other"
                          ]
                        },
                        "mainText": { "type": "string" },
                        "username": { "type": "string" },
                        "timestampText": { "type": "string" }
                      }
                    },
                    "triangulation": {
                      "type": "object",
                      "properties": {
                        "targetUrl": { "type": "string", "format": "uri" },
                        "queries": {
                          "type": "array",
                          "items": { "type": "string" }
                        }
                      }
                    },
                    "audit": {
                      "type": "object",
                      "properties": {
                        "originScore": { "type": "number" },
                        "temporalScore": { "type": "number" },
                        "totalScore": {
                          "type": "number",
                          "description": "Postcard score (0-100)"
                        },
                        "auditLog": {
                          "type": "array",
                          "items": { "type": "string" }
                        }
                      }
                    },
                    "corroboration": {
                      "$ref": "#/components/schemas/Corroboration"
                    },
                    "timestamp": { "type": "string", "format": "date-time" },
                    "postcardId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "302": {
            "description": "Redirect to UI (if Accept header is not application/json)"
          },
          "400": {
            "description": "Missing or invalid URL parameter"
          },
          "404": {
            "description": "Analysis not found"
          }
        }
      },
      "post": {
        "summary": "Initiate new trace",
        "description": "Submits a new URL for forensic verification. Returns a postcard ID for polling status.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostcardRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Postcard analysis started, use postcardId to poll for status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "postcardId": { "type": "string", "format": "uuid" },
                    "status": { "type": "string", "enum": ["processing"] },
                    "message": { "type": "string" }
                  }
                },
                "example": {
                  "postcardId": "550e8400-e29b-41d4-a716-446655440000",
                  "status": "processing",
                  "message": "Analysis started"
                }
              }
            }
          },
          "200": {
            "description": "Existing completed analysis returned (cache hit)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostcardReport"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PostcardRequest": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Direct URL of the social media post"
          },
          "userApiKey": {
            "type": "string",
            "description": "Optional API key for external integrations"
          },
          "forceRefresh": {
            "type": "boolean",
            "default": false,
            "description": "Bypass cache and perform a fresh audit"
          }
        },
        "required": ["url"]
      },
      "PostcardReport": {
        "type": "object",
        "properties": {
          "postcard": {
            "type": "object",
            "properties": {
              "platform": {
                "type": "string",
                "enum": ["X", "YouTube", "Reddit", "Instagram", "Other"]
              },
              "mainText": {
                "type": "string"
              },
              "username": {
                "type": "string"
              },
              "timestampText": {
                "type": "string"
              }
            }
          },
          "markdown": {
            "type": "string",
            "description": "Full extracted content in Markdown format"
          },
          "triangulation": {
            "type": "object",
            "properties": {
              "targetUrl": {
                "type": "string",
                "format": "uri"
              },
              "queries": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "audit": {
            "type": "object",
            "properties": {
              "originScore": {
                "type": "number"
              },
              "temporalScore": {
                "type": "number"
              },
              "totalScore": {
                "type": "number",
                "description": "Postcard score (0-100)"
              },
              "auditLog": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "corroboration": {
            "$ref": "#/components/schemas/Corroboration"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "analysisId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "Corroboration": {
        "type": "object",
        "properties": {
          "primarySources": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "url": {
                  "type": "string",
                  "format": "uri"
                },
                "title": {
                  "type": "string"
                },
                "source": {
                  "type": "string"
                },
                "snippet": {
                  "type": "string"
                },
                "relevance": {
                  "type": "string",
                  "enum": ["supporting", "refuting", "neutral"]
                }
              }
            }
          },
          "queriesExecuted": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "query": {
                  "type": "string"
                },
                "sourcesFound": {
                  "type": "number"
                }
              }
            }
          },
          "verdict": {
            "type": "string",
            "enum": [
              "verified",
              "disputed",
              "inconclusive",
              "insufficient_data"
            ]
          },
          "summary": {
            "type": "string"
          },
          "confidenceScore": {
            "type": "number"
          },
          "corroborationLog": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}